+++ addons/test_mail/tests/__init__.py
@@ -3,6 +3,7 @@
 from . import test_invite
 from . import test_ir_actions
 from . import test_mail_activity
+from . import test_mail_alias
 from . import test_mail_composer
 from . import test_mail_composer_mixin
 from . import test_mail_followers
